% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_db2_not_in_db1.R
\name{find_nots}
\alias{find_nots}
\title{Merge elements present only in one of the two databases}
\usage{
find_nots(df1, df2, folder = NA)
}
\arguments{
\item{df1}{database from mefu}

\item{df2}{database from comixtime}
}
\value{
a dataframe
}
\description{
Merge elements present only in one of the two databases
}
