% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import_df}
\alias{import_df}
\title{Import a dataframe from an url}
\usage{
import_df(address)
}
\arguments{
\item{address}{address from which you want to download the database from}
}
\description{
Import a dataframe from an url
}
