% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_db2_not_in_db1.R
\name{in_db1_not_in_db2}
\alias{in_db1_not_in_db2}
\title{Find elements present in df1 but not in df2}
\usage{
in_db1_not_in_db2(df1, df2, folder = NA)
}
\arguments{
\item{df1}{database from mefu}

\item{df2}{database from comixtime}
}
\value{
a dataframe
}
\description{
Find elements present in df1 but not in df2
}
