% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_db2_not_in_db1.R
\name{in_db2_not_in_db1}
\alias{in_db2_not_in_db1}
\title{Find elements present in df2 but not in df1}
\usage{
in_db2_not_in_db1(df1, df2, folder = NA)
}
\arguments{
\item{df1}{database from mefu}

\item{df2}{database from comixtime}

\item{folder}{if set to "choose" let you choose navigating in your folder system.
If a path is provided it'll save there}
}
\value{
a dataframe
}
\description{
Find elements present in df2 but not in df1
}
