% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_databases.R
\name{in_merged_and_in_db2}
\alias{in_merged_and_in_db2}
\title{Find elements present both in df1 and in df2}
\usage{
in_merged_and_in_db2(df1, df2, folder = NA)
}
\arguments{
\item{df1}{database from mefu}

\item{df2}{database from comixtime}

\item{folder}{if set to "choose" let you choose navigating in your folder system.
If a path is provided it'll save there. N.B.: The path format has to be of
the kind "C://folder1//folder2"}
}
\value{
a dataframe
}
\description{
Create the merged database, then add all the unmerged data from
df2
}
