% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lista_editori_unici}
\alias{lista_editori_unici}
\title{Lists all the unique instances of "Editori" present in a data frame}
\usage{
lista_editori_unici(df)
}
\arguments{
\item{df}{a dataframe}
}
\description{
Lists all the unique instances of "Editori" present in a data frame
}
