% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lista_insegna_presso_unici}
\alias{lista_insegna_presso_unici}
\title{Lists all the unique instances of "Insegna.presso" present in a data frame}
\usage{
lista_insegna_presso_unici(df)
}
\arguments{
\item{df}{a dataframe}
}
\description{
Lists all the unique instances of "Insegna.presso" present in a data frame
}
