% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lista_isbn_unici}
\alias{lista_isbn_unici}
\title{Lists all the unique instances of "ISBN" present in a data frame}
\usage{
lista_isbn_unici(df)
}
\arguments{
\item{df}{a dataframe}
}
\description{
Lists all the unique instances of "ISBN" present in a data frame
}
