% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lista_prima_pubblicazione_unici}
\alias{lista_prima_pubblicazione_unici}
\title{Lists all the unique instances of "Paesi.di.prima.pubblicazione" present in a data frame}
\usage{
lista_prima_pubblicazione_unici(df)
}
\arguments{
\item{df}{a dataframe}
}
\description{
Lists all the unique instances of "Paesi.di.prima.pubblicazione" present in a data frame
}
