% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_databases.R
\name{match_db_intersections_with_dbmefu}
\alias{match_db_intersections_with_dbmefu}
\title{Merge elements present only in one of the two databases}
\usage{
match_db_intersections_with_dbmefu(filename)
}
\arguments{
\item{filename}{a filename}
}
\value{
a dataframe
}
\description{
Merge elements present only in one of the two databases
}
