% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_db.R
\name{merge_db}
\alias{merge_db}
\title{Merge databases}
\usage{
merge_db(df1, df2, folder = NA)
}
\arguments{
\item{df1}{database from mefu}

\item{df2}{database from comixtime}

\item{folder}{if set to "choose" let you choose navigating in your folder system.
If a path is provided it'll save there}
}
\value{
a merged database
}
\description{
Merges df1 and df2
}
