% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulizia.R
\name{ripulisci_df}
\alias{ripulisci_df}
\title{Clean the dataframe so you can work with it}
\usage{
ripulisci_df(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a clean dataframe
}
\description{
Clean the dataframe so you can work with it
}
