% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all.R
\name{run_all}
\alias{run_all}
\title{Creates all the dataframes and save them in the specified folder}
\usage{
run_all(
 
    address1 = "https://docs.google.com/spreadsheets/d/1wKeDUhL4TJJ9yVUDN49kOANU_5Y77r6-vIxaq4Wgdks/",
 
    address2 = "https://docs.google.com/spreadsheets/d/1AipazAj6Ebfuv0Kek4xMRYva8sodUQs-4AI3li0IbzY/",
  folder
)
}
\arguments{
\item{address1}{address from which you want to download the df1 from}

\item{address2}{address from which you want to download the df2 from}

\item{folder}{if set to "choose" let you choose navigating in your folder system.
If a path is provided it'll save there. N.B.: The path format has to be of
the kind "C://folder1//folder2"}
}
\description{
Creates all the dataframes and save them in the specified folder
}
