% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{save_df}
\alias{save_df}
\title{Save a data frame in the specified folder}
\usage{
save_df(df, filename, folder)
}
\arguments{
\item{df}{a dataframe}

\item{filename}{a filename}

\item{folder}{if set to "choose" let you choose navigating in your folder system.
If a path is provided it'll save there. N.B.: The path format has to be of
the kind "C://folder1//folder2"}
}
\description{
Save a data frame in the specified folder
}
