% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_param_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  address,
  address1,
  address2,
  colonna,
  dbmefu,
  df,
  df1,
  df2,
  dfmerged,
  filename,
  folder,
  names
)
}
\arguments{
\item{address}{address from which you want to download the database from}

\item{address1}{address from which you want to download the df1 from}

\item{address2}{address from which you want to download the df2 from}

\item{colonna}{a column}

\item{dbmefu}{the one and only MeFu database}

\item{df}{a dataframe}

\item{df1}{database from mefu}

\item{df2}{database from comixtime}

\item{dfmerged}{a merged data frame}

\item{filename}{a filename}

\item{folder}{if set to "choose" let you choose navigating in your folder system.
If a path is provided it'll save there. N.B.: The path format has to be of
the kind "C://folder1//folder2"}

\item{names}{a list of names}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
