% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_param_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(df1, df2, folder)
}
\arguments{
\item{df1}{database from mefu}

\item{df2}{database from comixtime}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
